// src/ui/planet/renderer/solar-system-overlay/types.ts
export interface AsteroidFieldData {
	mesh: THREE.InstancedMesh
	count: number
	angles: Float32Array
	radii: Float32Array
	zOffsets: Float32Array
	scales: Float32Array
	rotationAxes: THREE.Vector3[]
	rotationSpeeds: Float32Array
}